Rewrite envelope editor as editor component, separate state per map
Move all envelope editor rendering, envelope zoom handling and envelope editor popups to new `CEnvelopeEditor` component in `game/editor/envelope_editor.cpp/h`.

Add separate `CEditorEditor::CState` class to encapsulate the state that is stored separately for each map.

Avoid using any non-const static variables for the envelope editor.
